LayoutTests/imported/w3c/web-platform-tests/encoding/single-byte-decoder-expected.txt


FAIL ibm866: 866 (XMLHttpRequest) assert_equals: ibm866:26 expected 26 but got 28
FAIL ibm866: 866 (TextDecoder) Can't find variable: TextDecoder
PASS ibm866: 866 (document.characterSet and document.inputEncoding) 
FAIL ibm866: cp866 (XMLHttpRequest) assert_equals: ibm866:26 expected 26 but got 28
FAIL ibm866: cp866 (TextDecoder) Can't find variable: TextDecoder
PASS ibm866: cp866 (document.characterSet and document.inputEncoding) 
FAIL ibm866: csibm866 (XMLHttpRequest) assert_equals: ibm866:26 expected 26 but got 28
FAIL ibm866: csibm866 (TextDecoder) Can't find variable: TextDecoder
PASS ibm866: csibm866 (document.characterSet and document.inputEncoding) 
FAIL ibm866: ibm866 (XMLHttpRequest) assert_equals: ibm866:26 expected 26 but got 28
FAIL ibm866: ibm866 (TextDecoder) Can't find variable: TextDecoder
PASS ibm866: ibm866 (document.characterSet and document.inputEncoding) 
PASS iso-8859-2: csisolatin2 (XMLHttpRequest) 
FAIL iso-8859-2: csisolatin2 (TextDecoder) Can't find variable: TextDecoder
PASS iso-8859-2: csisolatin2 (document.characterSet and document.inputEncoding) 
PASS iso-8859-2: iso-8859-2 (XMLHttpRequest) 
FAIL iso-8859-2: iso-8859-2 (TextDecoder) Can't find variable: TextDecoder
PASS iso-8859-2: iso-8859-2 (document.characterSet and document.inputEncoding) 
PASS iso-8859-2: iso-ir-101 (XMLHttpRequest) 
FAIL iso-8859-2: iso-ir-101 (TextDecoder) Can't find variable: TextDecoder
PASS iso-8859-2: iso-ir-101 (document.characterSet and document.inputEncoding) 
PASS iso-8859-2: iso8859-2 (XMLHttpRequest) 
FAIL iso-8859-2: iso8859-2 (TextDecoder) Can't find variable: TextDecoder
PASS iso-8859-2: iso8859-2 (document.characterSet and document.inputEncoding) 
PASS iso-8859-2: iso88592 (XMLHttpRequest) 
FAIL iso-8859-2: iso88592 (TextDecoder) Can't find variable: TextDecoder
PASS iso-8859-2: iso88592 (document.characterSet and document.inputEncoding) 
PASS iso-8859-2: iso_8859-2 (XMLHttpRequest) 
FAIL iso-8859-2: iso_8859-2 (TextDecoder) Can't find variable: TextDecoder
PASS iso-8859-2: iso_8859-2 (document.characterSet and document.inputEncoding) 
PASS iso-8859-2: iso_8859-2:1987 (XMLHttpRequest) 
FAIL iso-8859-2: iso_8859-2:1987 (TextDecoder) Can't find variable: TextDecoder
PASS iso-8859-2: iso_8859-2:1987 (document.characterSet and document.inputEncoding) 
PASS iso-8859-2: l2 (XMLHttpRequest) 
FAIL iso-8859-2: l2 (TextDecoder) Can't find variable: TextDecoder
PASS iso-8859-2: l2 (document.characterSet and document.inputEncoding) 
PASS iso-8859-2: latin2 (XMLHttpRequest) 
FAIL iso-8859-2: latin2 (TextDecoder) Can't find variable: TextDecoder
PASS iso-8859-2: latin2 (document.characterSet and document.inputEncoding) 
PASS iso-8859-3: csisolatin3 (XMLHttpRequest) 
FAIL iso-8859-3: csisolatin3 (TextDecoder) Can't find variable: TextDecoder
PASS iso-8859-3: csisolatin3 (document.characterSet and document.inputEncoding) 
PASS iso-8859-3: iso-8859-3 (XMLHttpRequest) 
FAIL iso-8859-3: iso-8859-3 (TextDecoder) Can't find variable: TextDecoder
PASS iso-8859-3: iso-8859-3 (document.characterSet and document.inputEncoding) 
PASS iso-8859-3: iso-ir-109 (XMLHttpRequest) 
FAIL iso-8859-3: iso-ir-109 (TextDecoder) Can't find variable: TextDecoder
PASS iso-8859-3: iso-ir-109 (document.characterSet and document.inputEncoding) 
PASS iso-8859-3: iso8859-3 (XMLHttpRequest) 
FAIL iso-8859-3: iso8859-3 (TextDecoder) Can't find variable: TextDecoder
PASS iso-8859-3: iso8859-3 (document.characterSet and document.inputEncoding) 
PASS iso-8859-3: iso88593 (XMLHttpRequest) 
FAIL iso-8859-3: iso88593 (TextDecoder) Can't find variable: TextDecoder
PASS iso-8859-3: iso88593 (document.characterSet and document.inputEncoding) 
PASS iso-8859-3: iso_8859-3 (XMLHttpRequest) 
FAIL iso-8859-3: iso_8859-3 (TextDecoder) Can't find variable: TextDecoder
PASS iso-8859-3: iso_8859-3 (document.characterSet and document.inputEncoding) 
PASS iso-8859-3: iso_8859-3:1988 (XMLHttpRequest) 
FAIL iso-8859-3: iso_8859-3:1988 (TextDecoder) Can't find variable: TextDecoder
PASS iso-8859-3: iso_8859-3:1988 (document.characterSet and document.inputEncoding) 
PASS iso-8859-3: l3 (XMLHttpRequest) 
FAIL iso-8859-3: l3 (TextDecoder) Can't find variable: TextDecoder
PASS iso-8859-3: l3 (document.characterSet and document.inputEncoding) 
PASS iso-8859-3: latin3 (XMLHttpRequest) 
FAIL iso-8859-3: latin3 (TextDecoder) Can't find variable: TextDecoder
PASS iso-8859-3: latin3 (document.characterSet and document.inputEncoding) 
PASS iso-8859-4: csisolatin4 (XMLHttpRequest) 
FAIL iso-8859-4: csisolatin4 (TextDecoder) Can't find variable: TextDecoder
PASS iso-8859-4: csisolatin4 (document.characterSet and document.inputEncoding) 
PASS iso-8859-4: iso-8859-4 (XMLHttpRequest) 
FAIL iso-8859-4: iso-8859-4 (TextDecoder) Can't find variable: TextDecoder
PASS iso-8859-4: iso-8859-4 (document.characterSet and document.inputEncoding) 
PASS iso-8859-4: iso-ir-110 (XMLHttpRequest) 
FAIL iso-8859-4: iso-ir-110 (TextDecoder) Can't find variable: TextDecoder
PASS iso-8859-4: iso-ir-110 (document.characterSet and document.inputEncoding) 
PASS iso-8859-4: iso8859-4 (XMLHttpRequest) 
FAIL iso-8859-4: iso8859-4 (TextDecoder) Can't find variable: TextDecoder
PASS iso-8859-4: iso8859-4 (document.characterSet and document.inputEncoding) 
PASS iso-8859-4: iso88594 (XMLHttpRequest) 
FAIL iso-8859-4: iso88594 (TextDecoder) Can't find variable: TextDecoder
PASS iso-8859-4: iso88594 (document.characterSet and document.inputEncoding) 
PASS iso-8859-4: iso_8859-4 (XMLHttpRequest) 
FAIL iso-8859-4: iso_8859-4 (TextDecoder) Can't find variable: TextDecoder
PASS iso-8859-4: iso_8859-4 (document.characterSet and document.inputEncoding) 
PASS iso-8859-4: iso_8859-4:1988 (XMLHttpRequest) 
FAIL iso-8859-4: iso_8859-4:1988 (TextDecoder) Can't find variable: TextDecoder
PASS iso-8859-4: iso_8859-4:1988 (document.characterSet and document.inputEncoding) 
PASS iso-8859-4: l4 (XMLHttpRequest) 
FAIL iso-8859-4: l4 (TextDecoder) Can't find variable: TextDecoder
PASS iso-8859-4: l4 (document.characterSet and document.inputEncoding) 
PASS iso-8859-4: latin4 (XMLHttpRequest) 
FAIL iso-8859-4: latin4 (TextDecoder) Can't find variable: TextDecoder
PASS iso-8859-4: latin4 (document.characterSet and document.inputEncoding) 
PASS iso-8859-5: csisolatincyrillic (XMLHttpRequest) 
FAIL iso-8859-5: csisolatincyrillic (TextDecoder) Can't find variable: TextDecoder
PASS iso-8859-5: csisolatincyrillic (document.characterSet and document.inputEncoding) 
PASS iso-8859-5: cyrillic (XMLHttpRequest) 
FAIL iso-8859-5: cyrillic (TextDecoder) Can't find variable: TextDecoder
PASS iso-8859-5: cyrillic (document.characterSet and document.inputEncoding) 
PASS iso-8859-5: iso-8859-5 (XMLHttpRequest) 
FAIL iso-8859-5: iso-8859-5 (TextDecoder) Can't find variable: TextDecoder
PASS iso-8859-5: iso-8859-5 (document.characterSet and document.inputEncoding) 
PASS iso-8859-5: iso-ir-144 (XMLHttpRequest) 
FAIL iso-8859-5: iso-ir-144 (TextDecoder) Can't find variable: TextDecoder
PASS iso-8859-5: iso-ir-144 (document.characterSet and document.inputEncoding) 
PASS iso-8859-5: iso8859-5 (XMLHttpRequest) 
FAIL iso-8859-5: iso8859-5 (TextDecoder) Can't find variable: TextDecoder
PASS iso-8859-5: iso8859-5 (document.characterSet and document.inputEncoding) 
PASS iso-8859-5: iso88595 (XMLHttpRequest) 
FAIL iso-8859-5: iso88595 (TextDecoder) Can't find variable: TextDecoder
PASS iso-8859-5: iso88595 (document.characterSet and document.inputEncoding) 
PASS iso-8859-5: iso_8859-5 (XMLHttpRequest) 
FAIL iso-8859-5: iso_8859-5 (TextDecoder) Can't find variable: TextDecoder
PASS iso-8859-5: iso_8859-5 (document.characterSet and document.inputEncoding) 
PASS iso-8859-5: iso_8859-5:1988 (XMLHttpRequest) 
FAIL iso-8859-5: iso_8859-5:1988 (TextDecoder) Can't find variable: TextDecoder
PASS iso-8859-5: iso_8859-5:1988 (document.characterSet and document.inputEncoding) 
PASS iso-8859-6: arabic (XMLHttpRequest) 
FAIL iso-8859-6: arabic (TextDecoder) Can't find variable: TextDecoder
PASS iso-8859-6: arabic (document.characterSet and document.inputEncoding) 
PASS iso-8859-6: asmo-708 (XMLHttpRequest) 
FAIL iso-8859-6: asmo-708 (TextDecoder) Can't find variable: TextDecoder
PASS iso-8859-6: asmo-708 (document.characterSet and document.inputEncoding) 
PASS iso-8859-6: csiso88596e (XMLHttpRequest) 
FAIL iso-8859-6: csiso88596e (TextDecoder) Can't find variable: TextDecoder
PASS iso-8859-6: csiso88596e (document.characterSet and document.inputEncoding) 
PASS iso-8859-6: csiso88596i (XMLHttpRequest) 
FAIL iso-8859-6: csiso88596i (TextDecoder) Can't find variable: TextDecoder
PASS iso-8859-6: csiso88596i (document.characterSet and document.inputEncoding) 
PASS iso-8859-6: csisolatinarabic (XMLHttpRequest) 
FAIL iso-8859-6: csisolatinarabic (TextDecoder) Can't find variable: TextDecoder
PASS iso-8859-6: csisolatinarabic (document.characterSet and document.inputEncoding) 
PASS iso-8859-6: ecma-114 (XMLHttpRequest) 
FAIL iso-8859-6: ecma-114 (TextDecoder) Can't find variable: TextDecoder
PASS iso-8859-6: ecma-114 (document.characterSet and document.inputEncoding) 
PASS iso-8859-6: iso-8859-6 (XMLHttpRequest) 
FAIL iso-8859-6: iso-8859-6 (TextDecoder) Can't find variable: TextDecoder
PASS iso-8859-6: iso-8859-6 (document.characterSet and document.inputEncoding) 
PASS iso-8859-6: iso-8859-6-e (XMLHttpRequest) 
FAIL iso-8859-6: iso-8859-6-e (TextDecoder) Can't find variable: TextDecoder
PASS iso-8859-6: iso-8859-6-e (document.characterSet and document.inputEncoding) 
PASS iso-8859-6: iso-8859-6-i (XMLHttpRequest) 
FAIL iso-8859-6: iso-8859-6-i (TextDecoder) Can't find variable: TextDecoder
PASS iso-8859-6: iso-8859-6-i (document.characterSet and document.inputEncoding) 
PASS iso-8859-6: iso-ir-127 (XMLHttpRequest) 
FAIL iso-8859-6: iso-ir-127 (TextDecoder) Can't find variable: TextDecoder
PASS iso-8859-6: iso-ir-127 (document.characterSet and document.inputEncoding) 
PASS iso-8859-6: iso8859-6 (XMLHttpRequest) 
FAIL iso-8859-6: iso8859-6 (TextDecoder) Can't find variable: TextDecoder
PASS iso-8859-6: iso8859-6 (document.characterSet and document.inputEncoding) 
PASS iso-8859-6: iso88596 (XMLHttpRequest) 
FAIL iso-8859-6: iso88596 (TextDecoder) Can't find variable: TextDecoder
PASS iso-8859-6: iso88596 (document.characterSet and document.inputEncoding) 
PASS iso-8859-6: iso_8859-6 (XMLHttpRequest) 
FAIL iso-8859-6: iso_8859-6 (TextDecoder) Can't find variable: TextDecoder
PASS iso-8859-6: iso_8859-6 (document.characterSet and document.inputEncoding) 
PASS iso-8859-6: iso_8859-6:1987 (XMLHttpRequest) 
FAIL iso-8859-6: iso_8859-6:1987 (TextDecoder) Can't find variable: TextDecoder
PASS iso-8859-6: iso_8859-6:1987 (document.characterSet and document.inputEncoding) 
PASS iso-8859-7: csisolatingreek (XMLHttpRequest) 
FAIL iso-8859-7: csisolatingreek (TextDecoder) Can't find variable: TextDecoder
PASS iso-8859-7: csisolatingreek (document.characterSet and document.inputEncoding) 
PASS iso-8859-7: ecma-118 (XMLHttpRequest) 
FAIL iso-8859-7: ecma-118 (TextDecoder) Can't find variable: TextDecoder
PASS iso-8859-7: ecma-118 (document.characterSet and document.inputEncoding) 
PASS iso-8859-7: elot_928 (XMLHttpRequest) 
FAIL iso-8859-7: elot_928 (TextDecoder) Can't find variable: TextDecoder
PASS iso-8859-7: elot_928 (document.characterSet and document.inputEncoding) 
PASS iso-8859-7: greek (XMLHttpRequest) 
FAIL iso-8859-7: greek (TextDecoder) Can't find variable: TextDecoder
PASS iso-8859-7: greek (document.characterSet and document.inputEncoding) 
PASS iso-8859-7: greek8 (XMLHttpRequest) 
FAIL iso-8859-7: greek8 (TextDecoder) Can't find variable: TextDecoder
PASS iso-8859-7: greek8 (document.characterSet and document.inputEncoding) 
PASS iso-8859-7: iso-8859-7 (XMLHttpRequest) 
FAIL iso-8859-7: iso-8859-7 (TextDecoder) Can't find variable: TextDecoder
PASS iso-8859-7: iso-8859-7 (document.characterSet and document.inputEncoding) 
PASS iso-8859-7: iso-ir-126 (XMLHttpRequest) 
FAIL iso-8859-7: iso-ir-126 (TextDecoder) Can't find variable: TextDecoder
PASS iso-8859-7: iso-ir-126 (document.characterSet and document.inputEncoding) 
PASS iso-8859-7: iso8859-7 (XMLHttpRequest) 
FAIL iso-8859-7: iso8859-7 (TextDecoder) Can't find variable: TextDecoder
PASS iso-8859-7: iso8859-7 (document.characterSet and document.inputEncoding) 
PASS iso-8859-7: iso88597 (XMLHttpRequest) 
FAIL iso-8859-7: iso88597 (TextDecoder) Can't find variable: TextDecoder
PASS iso-8859-7: iso88597 (document.characterSet and document.inputEncoding) 
PASS iso-8859-7: iso_8859-7 (XMLHttpRequest) 
FAIL iso-8859-7: iso_8859-7 (TextDecoder) Can't find variable: TextDecoder
PASS iso-8859-7: iso_8859-7 (document.characterSet and document.inputEncoding) 
PASS iso-8859-7: iso_8859-7:1987 (XMLHttpRequest) 
FAIL iso-8859-7: iso_8859-7:1987 (TextDecoder) Can't find variable: TextDecoder
PASS iso-8859-7: iso_8859-7:1987 (document.characterSet and document.inputEncoding) 
PASS iso-8859-7: sun_eu_greek (XMLHttpRequest) 
FAIL iso-8859-7: sun_eu_greek (TextDecoder) Can't find variable: TextDecoder
PASS iso-8859-7: sun_eu_greek (document.characterSet and document.inputEncoding) 
PASS iso-8859-8: csiso88598e (XMLHttpRequest) 
FAIL iso-8859-8: csiso88598e (TextDecoder) Can't find variable: TextDecoder
PASS iso-8859-8: csiso88598e (document.characterSet and document.inputEncoding) 
PASS iso-8859-8: csisolatinhebrew (XMLHttpRequest) 
FAIL iso-8859-8: csisolatinhebrew (TextDecoder) Can't find variable: TextDecoder
PASS iso-8859-8: csisolatinhebrew (document.characterSet and document.inputEncoding) 
PASS iso-8859-8: hebrew (XMLHttpRequest) 
FAIL iso-8859-8: hebrew (TextDecoder) Can't find variable: TextDecoder
PASS iso-8859-8: hebrew (document.characterSet and document.inputEncoding) 
PASS iso-8859-8: iso-8859-8 (XMLHttpRequest) 
FAIL iso-8859-8: iso-8859-8 (TextDecoder) Can't find variable: TextDecoder
PASS iso-8859-8: iso-8859-8 (document.characterSet and document.inputEncoding) 
PASS iso-8859-8: iso-8859-8-e (XMLHttpRequest) 
FAIL iso-8859-8: iso-8859-8-e (TextDecoder) Can't find variable: TextDecoder
PASS iso-8859-8: iso-8859-8-e (document.characterSet and document.inputEncoding) 
PASS iso-8859-8: iso-ir-138 (XMLHttpRequest) 
FAIL iso-8859-8: iso-ir-138 (TextDecoder) Can't find variable: TextDecoder
PASS iso-8859-8: iso-ir-138 (document.characterSet and document.inputEncoding) 
PASS iso-8859-8: iso8859-8 (XMLHttpRequest) 
FAIL iso-8859-8: iso8859-8 (TextDecoder) Can't find variable: TextDecoder
PASS iso-8859-8: iso8859-8 (document.characterSet and document.inputEncoding) 
PASS iso-8859-8: iso88598 (XMLHttpRequest) 
FAIL iso-8859-8: iso88598 (TextDecoder) Can't find variable: TextDecoder
PASS iso-8859-8: iso88598 (document.characterSet and document.inputEncoding) 
PASS iso-8859-8: iso_8859-8 (XMLHttpRequest) 
FAIL iso-8859-8: iso_8859-8 (TextDecoder) Can't find variable: TextDecoder
PASS iso-8859-8: iso_8859-8 (document.characterSet and document.inputEncoding) 
PASS iso-8859-8: iso_8859-8:1988 (XMLHttpRequest) 
FAIL iso-8859-8: iso_8859-8:1988 (TextDecoder) Can't find variable: TextDecoder
PASS iso-8859-8: iso_8859-8:1988 (document.characterSet and document.inputEncoding) 
PASS iso-8859-8: visual (XMLHttpRequest) 
FAIL iso-8859-8: visual (TextDecoder) Can't find variable: TextDecoder
PASS iso-8859-8: visual (document.characterSet and document.inputEncoding) 
PASS iso-8859-8-i: csiso88598i (XMLHttpRequest) 
FAIL iso-8859-8-i: csiso88598i (TextDecoder) Can't find variable: TextDecoder
PASS iso-8859-8-i: csiso88598i (document.characterSet and document.inputEncoding) 
PASS iso-8859-8-i: iso-8859-8-i (XMLHttpRequest) 
FAIL iso-8859-8-i: iso-8859-8-i (TextDecoder) Can't find variable: TextDecoder
PASS iso-8859-8-i: iso-8859-8-i (document.characterSet and document.inputEncoding) 
PASS iso-8859-8-i: logical (XMLHttpRequest) 
FAIL iso-8859-8-i: logical (TextDecoder) Can't find variable: TextDecoder
PASS iso-8859-8-i: logical (document.characterSet and document.inputEncoding) 
PASS iso-8859-10: csisolatin6 (XMLHttpRequest) 
FAIL iso-8859-10: csisolatin6 (TextDecoder) Can't find variable: TextDecoder
PASS iso-8859-10: csisolatin6 (document.characterSet and document.inputEncoding) 
PASS iso-8859-10: iso-8859-10 (XMLHttpRequest) 
FAIL iso-8859-10: iso-8859-10 (TextDecoder) Can't find variable: TextDecoder
PASS iso-8859-10: iso-8859-10 (document.characterSet and document.inputEncoding) 
PASS iso-8859-10: iso-ir-157 (XMLHttpRequest) 
FAIL iso-8859-10: iso-ir-157 (TextDecoder) Can't find variable: TextDecoder
PASS iso-8859-10: iso-ir-157 (document.characterSet and document.inputEncoding) 
PASS iso-8859-10: iso8859-10 (XMLHttpRequest) 
FAIL iso-8859-10: iso8859-10 (TextDecoder) Can't find variable: TextDecoder
PASS iso-8859-10: iso8859-10 (document.characterSet and document.inputEncoding) 
PASS iso-8859-10: iso885910 (XMLHttpRequest) 
FAIL iso-8859-10: iso885910 (TextDecoder) Can't find variable: TextDecoder
PASS iso-8859-10: iso885910 (document.characterSet and document.inputEncoding) 
PASS iso-8859-10: l6 (XMLHttpRequest) 
FAIL iso-8859-10: l6 (TextDecoder) Can't find variable: TextDecoder
PASS iso-8859-10: l6 (document.characterSet and document.inputEncoding) 
PASS iso-8859-10: latin6 (XMLHttpRequest) 
FAIL iso-8859-10: latin6 (TextDecoder) Can't find variable: TextDecoder
PASS iso-8859-10: latin6 (document.characterSet and document.inputEncoding) 
PASS iso-8859-13: iso-8859-13 (XMLHttpRequest) 
FAIL iso-8859-13: iso-8859-13 (TextDecoder) Can't find variable: TextDecoder
PASS iso-8859-13: iso-8859-13 (document.characterSet and document.inputEncoding) 
PASS iso-8859-13: iso8859-13 (XMLHttpRequest) 
FAIL iso-8859-13: iso8859-13 (TextDecoder) Can't find variable: TextDecoder
PASS iso-8859-13: iso8859-13 (document.characterSet and document.inputEncoding) 
PASS iso-8859-13: iso885913 (XMLHttpRequest) 
FAIL iso-8859-13: iso885913 (TextDecoder) Can't find variable: TextDecoder
PASS iso-8859-13: iso885913 (document.characterSet and document.inputEncoding) 
PASS iso-8859-14: iso-8859-14 (XMLHttpRequest) 
FAIL iso-8859-14: iso-8859-14 (TextDecoder) Can't find variable: TextDecoder
PASS iso-8859-14: iso-8859-14 (document.characterSet and document.inputEncoding) 
PASS iso-8859-14: iso8859-14 (XMLHttpRequest) 
FAIL iso-8859-14: iso8859-14 (TextDecoder) Can't find variable: TextDecoder
PASS iso-8859-14: iso8859-14 (document.characterSet and document.inputEncoding) 
PASS iso-8859-14: iso885914 (XMLHttpRequest) 
FAIL iso-8859-14: iso885914 (TextDecoder) Can't find variable: TextDecoder
PASS iso-8859-14: iso885914 (document.characterSet and document.inputEncoding) 
PASS iso-8859-15: csisolatin9 (XMLHttpRequest) 
FAIL iso-8859-15: csisolatin9 (TextDecoder) Can't find variable: TextDecoder
PASS iso-8859-15: csisolatin9 (document.characterSet and document.inputEncoding) 
PASS iso-8859-15: iso-8859-15 (XMLHttpRequest) 
FAIL iso-8859-15: iso-8859-15 (TextDecoder) Can't find variable: TextDecoder
PASS iso-8859-15: iso-8859-15 (document.characterSet and document.inputEncoding) 
PASS iso-8859-15: iso8859-15 (XMLHttpRequest) 
FAIL iso-8859-15: iso8859-15 (TextDecoder) Can't find variable: TextDecoder
PASS iso-8859-15: iso8859-15 (document.characterSet and document.inputEncoding) 
PASS iso-8859-15: iso885915 (XMLHttpRequest) 
FAIL iso-8859-15: iso885915 (TextDecoder) Can't find variable: TextDecoder
PASS iso-8859-15: iso885915 (document.characterSet and document.inputEncoding) 
PASS iso-8859-15: iso_8859-15 (XMLHttpRequest) 
FAIL iso-8859-15: iso_8859-15 (TextDecoder) Can't find variable: TextDecoder
PASS iso-8859-15: iso_8859-15 (document.characterSet and document.inputEncoding) 
PASS iso-8859-15: l9 (XMLHttpRequest) 
FAIL iso-8859-15: l9 (TextDecoder) Can't find variable: TextDecoder
PASS iso-8859-15: l9 (document.characterSet and document.inputEncoding) 
PASS iso-8859-16: iso-8859-16 (XMLHttpRequest) 
FAIL iso-8859-16: iso-8859-16 (TextDecoder) Can't find variable: TextDecoder
PASS iso-8859-16: iso-8859-16 (document.characterSet and document.inputEncoding) 
PASS koi8-r: cskoi8r (XMLHttpRequest) 
FAIL koi8-r: cskoi8r (TextDecoder) Can't find variable: TextDecoder
PASS koi8-r: cskoi8r (document.characterSet and document.inputEncoding) 
PASS koi8-r: koi (XMLHttpRequest) 
FAIL koi8-r: koi (TextDecoder) Can't find variable: TextDecoder
PASS koi8-r: koi (document.characterSet and document.inputEncoding) 
PASS koi8-r: koi8 (XMLHttpRequest) 
FAIL koi8-r: koi8 (TextDecoder) Can't find variable: TextDecoder
PASS koi8-r: koi8 (document.characterSet and document.inputEncoding) 
PASS koi8-r: koi8-r (XMLHttpRequest) 
FAIL koi8-r: koi8-r (TextDecoder) Can't find variable: TextDecoder
PASS koi8-r: koi8-r (document.characterSet and document.inputEncoding) 
PASS koi8-r: koi8_r (XMLHttpRequest) 
FAIL koi8-r: koi8_r (TextDecoder) Can't find variable: TextDecoder
PASS koi8-r: koi8_r (document.characterSet and document.inputEncoding) 
PASS koi8-u: koi8-u (XMLHttpRequest) 
FAIL koi8-u: koi8-u (TextDecoder) Can't find variable: TextDecoder
PASS koi8-u: koi8-u (document.characterSet and document.inputEncoding) 
PASS macintosh: csmacintosh (XMLHttpRequest) 
FAIL macintosh: csmacintosh (TextDecoder) Can't find variable: TextDecoder
PASS macintosh: csmacintosh (document.characterSet and document.inputEncoding) 
PASS macintosh: mac (XMLHttpRequest) 
FAIL macintosh: mac (TextDecoder) Can't find variable: TextDecoder
PASS macintosh: mac (document.characterSet and document.inputEncoding) 
PASS macintosh: macintosh (XMLHttpRequest) 
FAIL macintosh: macintosh (TextDecoder) Can't find variable: TextDecoder
PASS macintosh: macintosh (document.characterSet and document.inputEncoding) 
PASS macintosh: x-mac-roman (XMLHttpRequest) 
FAIL macintosh: x-mac-roman (TextDecoder) Can't find variable: TextDecoder
PASS macintosh: x-mac-roman (document.characterSet and document.inputEncoding) 
FAIL windows-874: dos-874 (XMLHttpRequest) assert_equals: windows-874:219 expected 65533 but got 63681
FAIL windows-874: dos-874 (TextDecoder) Can't find variable: TextDecoder
PASS windows-874: dos-874 (document.characterSet and document.inputEncoding) 
FAIL windows-874: iso-8859-11 (XMLHttpRequest) assert_equals: windows-874:219 expected 65533 but got 63681
FAIL windows-874: iso-8859-11 (TextDecoder) Can't find variable: TextDecoder
PASS windows-874: iso-8859-11 (document.characterSet and document.inputEncoding) 
FAIL windows-874: iso8859-11 (XMLHttpRequest) assert_equals: windows-874:219 expected 65533 but got 63681
FAIL windows-874: iso8859-11 (TextDecoder) Can't find variable: TextDecoder
PASS windows-874: iso8859-11 (document.characterSet and document.inputEncoding) 
FAIL windows-874: iso885911 (XMLHttpRequest) assert_equals: windows-874:219 expected 65533 but got 63681
FAIL windows-874: iso885911 (TextDecoder) Can't find variable: TextDecoder
PASS windows-874: iso885911 (document.characterSet and document.inputEncoding) 
FAIL windows-874: tis-620 (XMLHttpRequest) assert_equals: windows-874:219 expected 65533 but got 63681
FAIL windows-874: tis-620 (TextDecoder) Can't find variable: TextDecoder
PASS windows-874: tis-620 (document.characterSet and document.inputEncoding) 
FAIL windows-874: windows-874 (XMLHttpRequest) assert_equals: windows-874:219 expected 65533 but got 63681
FAIL windows-874: windows-874 (TextDecoder) Can't find variable: TextDecoder
PASS windows-874: windows-874 (document.characterSet and document.inputEncoding) 
PASS windows-1250: cp1250 (XMLHttpRequest) 
FAIL windows-1250: cp1250 (TextDecoder) Can't find variable: TextDecoder
PASS windows-1250: cp1250 (document.characterSet and document.inputEncoding) 
PASS windows-1250: windows-1250 (XMLHttpRequest) 
FAIL windows-1250: windows-1250 (TextDecoder) Can't find variable: TextDecoder
PASS windows-1250: windows-1250 (document.characterSet and document.inputEncoding) 
PASS windows-1250: x-cp1250 (XMLHttpRequest) 
FAIL windows-1250: x-cp1250 (TextDecoder) Can't find variable: TextDecoder
PASS windows-1250: x-cp1250 (document.characterSet and document.inputEncoding) 
PASS windows-1251: cp1251 (XMLHttpRequest) 
FAIL windows-1251: cp1251 (TextDecoder) Can't find variable: TextDecoder
PASS windows-1251: cp1251 (document.characterSet and document.inputEncoding) 
PASS windows-1251: windows-1251 (XMLHttpRequest) 
FAIL windows-1251: windows-1251 (TextDecoder) Can't find variable: TextDecoder
PASS windows-1251: windows-1251 (document.characterSet and document.inputEncoding) 
PASS windows-1251: x-cp1251 (XMLHttpRequest) 
FAIL windows-1251: x-cp1251 (TextDecoder) Can't find variable: TextDecoder
PASS windows-1251: x-cp1251 (document.characterSet and document.inputEncoding) 
PASS windows-1252: ansi_x3.4-1968 (XMLHttpRequest) 
FAIL windows-1252: ansi_x3.4-1968 (TextDecoder) Can't find variable: TextDecoder
PASS windows-1252: ansi_x3.4-1968 (document.characterSet and document.inputEncoding) 
PASS windows-1252: ascii (XMLHttpRequest) 
FAIL windows-1252: ascii (TextDecoder) Can't find variable: TextDecoder
PASS windows-1252: ascii (document.characterSet and document.inputEncoding) 
PASS windows-1252: cp1252 (XMLHttpRequest) 
FAIL windows-1252: cp1252 (TextDecoder) Can't find variable: TextDecoder
PASS windows-1252: cp1252 (document.characterSet and document.inputEncoding) 
PASS windows-1252: cp819 (XMLHttpRequest) 
FAIL windows-1252: cp819 (TextDecoder) Can't find variable: TextDecoder
PASS windows-1252: cp819 (document.characterSet and document.inputEncoding) 
PASS windows-1252: csisolatin1 (XMLHttpRequest) 
FAIL windows-1252: csisolatin1 (TextDecoder) Can't find variable: TextDecoder
PASS windows-1252: csisolatin1 (document.characterSet and document.inputEncoding) 
PASS windows-1252: ibm819 (XMLHttpRequest) 
FAIL windows-1252: ibm819 (TextDecoder) Can't find variable: TextDecoder
PASS windows-1252: ibm819 (document.characterSet and document.inputEncoding) 
PASS windows-1252: iso-8859-1 (XMLHttpRequest) 
FAIL windows-1252: iso-8859-1 (TextDecoder) Can't find variable: TextDecoder
PASS windows-1252: iso-8859-1 (document.characterSet and document.inputEncoding) 
PASS windows-1252: iso-ir-100 (XMLHttpRequest) 
FAIL windows-1252: iso-ir-100 (TextDecoder) Can't find variable: TextDecoder
PASS windows-1252: iso-ir-100 (document.characterSet and document.inputEncoding) 
PASS windows-1252: iso8859-1 (XMLHttpRequest) 
FAIL windows-1252: iso8859-1 (TextDecoder) Can't find variable: TextDecoder
PASS windows-1252: iso8859-1 (document.characterSet and document.inputEncoding) 
PASS windows-1252: iso88591 (XMLHttpRequest) 
FAIL windows-1252: iso88591 (TextDecoder) Can't find variable: TextDecoder
PASS windows-1252: iso88591 (document.characterSet and document.inputEncoding) 
PASS windows-1252: iso_8859-1 (XMLHttpRequest) 
FAIL windows-1252: iso_8859-1 (TextDecoder) Can't find variable: TextDecoder
PASS windows-1252: iso_8859-1 (document.characterSet and document.inputEncoding) 
PASS windows-1252: iso_8859-1:1987 (XMLHttpRequest) 
FAIL windows-1252: iso_8859-1:1987 (TextDecoder) Can't find variable: TextDecoder
PASS windows-1252: iso_8859-1:1987 (document.characterSet and document.inputEncoding) 
PASS windows-1252: l1 (XMLHttpRequest) 
FAIL windows-1252: l1 (TextDecoder) Can't find variable: TextDecoder
PASS windows-1252: l1 (document.characterSet and document.inputEncoding) 
PASS windows-1252: latin1 (XMLHttpRequest) 
FAIL windows-1252: latin1 (TextDecoder) Can't find variable: TextDecoder
PASS windows-1252: latin1 (document.characterSet and document.inputEncoding) 
PASS windows-1252: us-ascii (XMLHttpRequest) 
FAIL windows-1252: us-ascii (TextDecoder) Can't find variable: TextDecoder
PASS windows-1252: us-ascii (document.characterSet and document.inputEncoding) 
PASS windows-1252: windows-1252 (XMLHttpRequest) 
FAIL windows-1252: windows-1252 (TextDecoder) Can't find variable: TextDecoder
PASS windows-1252: windows-1252 (document.characterSet and document.inputEncoding) 
PASS windows-1252: x-cp1252 (XMLHttpRequest) 
FAIL windows-1252: x-cp1252 (TextDecoder) Can't find variable: TextDecoder
PASS windows-1252: x-cp1252 (document.characterSet and document.inputEncoding) 
FAIL windows-1253: cp1253 (XMLHttpRequest) assert_equals: windows-1253:170 expected 65533 but got 170
FAIL windows-1253: cp1253 (TextDecoder) Can't find variable: TextDecoder
PASS windows-1253: cp1253 (document.characterSet and document.inputEncoding) 
FAIL windows-1253: windows-1253 (XMLHttpRequest) assert_equals: windows-1253:170 expected 65533 but got 170
FAIL windows-1253: windows-1253 (TextDecoder) Can't find variable: TextDecoder
PASS windows-1253: windows-1253 (document.characterSet and document.inputEncoding) 
FAIL windows-1253: x-cp1253 (XMLHttpRequest) assert_equals: windows-1253:170 expected 65533 but got 170
FAIL windows-1253: x-cp1253 (TextDecoder) Can't find variable: TextDecoder
PASS windows-1253: x-cp1253 (document.characterSet and document.inputEncoding) 
PASS windows-1254: cp1254 (XMLHttpRequest) 
FAIL windows-1254: cp1254 (TextDecoder) Can't find variable: TextDecoder
PASS windows-1254: cp1254 (document.characterSet and document.inputEncoding) 
PASS windows-1254: csisolatin5 (XMLHttpRequest) 
FAIL windows-1254: csisolatin5 (TextDecoder) Can't find variable: TextDecoder
PASS windows-1254: csisolatin5 (document.characterSet and document.inputEncoding) 
PASS windows-1254: iso-8859-9 (XMLHttpRequest) 
FAIL windows-1254: iso-8859-9 (TextDecoder) Can't find variable: TextDecoder
PASS windows-1254: iso-8859-9 (document.characterSet and document.inputEncoding) 
PASS windows-1254: iso-ir-148 (XMLHttpRequest) 
FAIL windows-1254: iso-ir-148 (TextDecoder) Can't find variable: TextDecoder
PASS windows-1254: iso-ir-148 (document.characterSet and document.inputEncoding) 
PASS windows-1254: iso8859-9 (XMLHttpRequest) 
FAIL windows-1254: iso8859-9 (TextDecoder) Can't find variable: TextDecoder
PASS windows-1254: iso8859-9 (document.characterSet and document.inputEncoding) 
PASS windows-1254: iso88599 (XMLHttpRequest) 
FAIL windows-1254: iso88599 (TextDecoder) Can't find variable: TextDecoder
PASS windows-1254: iso88599 (document.characterSet and document.inputEncoding) 
PASS windows-1254: iso_8859-9 (XMLHttpRequest) 
FAIL windows-1254: iso_8859-9 (TextDecoder) Can't find variable: TextDecoder
PASS windows-1254: iso_8859-9 (document.characterSet and document.inputEncoding) 
PASS windows-1254: iso_8859-9:1989 (XMLHttpRequest) 
FAIL windows-1254: iso_8859-9:1989 (TextDecoder) Can't find variable: TextDecoder
PASS windows-1254: iso_8859-9:1989 (document.characterSet and document.inputEncoding) 
PASS windows-1254: l5 (XMLHttpRequest) 
FAIL windows-1254: l5 (TextDecoder) Can't find variable: TextDecoder
PASS windows-1254: l5 (document.characterSet and document.inputEncoding) 
PASS windows-1254: latin5 (XMLHttpRequest) 
FAIL windows-1254: latin5 (TextDecoder) Can't find variable: TextDecoder
PASS windows-1254: latin5 (document.characterSet and document.inputEncoding) 
PASS windows-1254: windows-1254 (XMLHttpRequest) 
FAIL windows-1254: windows-1254 (TextDecoder) Can't find variable: TextDecoder
PASS windows-1254: windows-1254 (document.characterSet and document.inputEncoding) 
PASS windows-1254: x-cp1254 (XMLHttpRequest) 
FAIL windows-1254: x-cp1254 (TextDecoder) Can't find variable: TextDecoder
PASS windows-1254: x-cp1254 (document.characterSet and document.inputEncoding) 
PASS windows-1255: cp1255 (XMLHttpRequest) 
FAIL windows-1255: cp1255 (TextDecoder) Can't find variable: TextDecoder
PASS windows-1255: cp1255 (document.characterSet and document.inputEncoding) 
PASS windows-1255: windows-1255 (XMLHttpRequest) 
FAIL windows-1255: windows-1255 (TextDecoder) Can't find variable: TextDecoder
PASS windows-1255: windows-1255 (document.characterSet and document.inputEncoding) 
PASS windows-1255: x-cp1255 (XMLHttpRequest) 
FAIL windows-1255: x-cp1255 (TextDecoder) Can't find variable: TextDecoder
PASS windows-1255: x-cp1255 (document.characterSet and document.inputEncoding) 
PASS windows-1256: cp1256 (XMLHttpRequest) 
FAIL windows-1256: cp1256 (TextDecoder) Can't find variable: TextDecoder
PASS windows-1256: cp1256 (document.characterSet and document.inputEncoding) 
PASS windows-1256: windows-1256 (XMLHttpRequest) 
FAIL windows-1256: windows-1256 (TextDecoder) Can't find variable: TextDecoder
PASS windows-1256: windows-1256 (document.characterSet and document.inputEncoding) 
PASS windows-1256: x-cp1256 (XMLHttpRequest) 
FAIL windows-1256: x-cp1256 (TextDecoder) Can't find variable: TextDecoder
PASS windows-1256: x-cp1256 (document.characterSet and document.inputEncoding) 
PASS windows-1257: cp1257 (XMLHttpRequest) 
FAIL windows-1257: cp1257 (TextDecoder) Can't find variable: TextDecoder
PASS windows-1257: cp1257 (document.characterSet and document.inputEncoding) 
PASS windows-1257: windows-1257 (XMLHttpRequest) 
FAIL windows-1257: windows-1257 (TextDecoder) Can't find variable: TextDecoder
PASS windows-1257: windows-1257 (document.characterSet and document.inputEncoding) 
PASS windows-1257: x-cp1257 (XMLHttpRequest) 
FAIL windows-1257: x-cp1257 (TextDecoder) Can't find variable: TextDecoder
PASS windows-1257: x-cp1257 (document.characterSet and document.inputEncoding) 
PASS windows-1258: cp1258 (XMLHttpRequest) 
FAIL windows-1258: cp1258 (TextDecoder) Can't find variable: TextDecoder
PASS windows-1258: cp1258 (document.characterSet and document.inputEncoding) 
PASS windows-1258: windows-1258 (XMLHttpRequest) 
FAIL windows-1258: windows-1258 (TextDecoder) Can't find variable: TextDecoder
PASS windows-1258: windows-1258 (document.characterSet and document.inputEncoding) 
PASS windows-1258: x-cp1258 (XMLHttpRequest) 
FAIL windows-1258: x-cp1258 (TextDecoder) Can't find variable: TextDecoder
PASS windows-1258: x-cp1258 (document.characterSet and document.inputEncoding) 
PASS x-mac-cyrillic: x-mac-cyrillic (XMLHttpRequest) 
FAIL x-mac-cyrillic: x-mac-cyrillic (TextDecoder) Can't find variable: TextDecoder
PASS x-mac-cyrillic: x-mac-cyrillic (document.characterSet and document.inputEncoding) 
PASS x-mac-cyrillic: x-mac-ukrainian (XMLHttpRequest) 
FAIL x-mac-cyrillic: x-mac-ukrainian (TextDecoder) Can't find variable: TextDecoder
PASS x-mac-cyrillic: x-mac-ukrainian (document.characterSet and document.inputEncoding) 
